server/requirements.txt
Flask>=0.12,<0.13
gunicorn>=19,<20
PyYAML>=3.13
redis>=3.0.1